Created progressbar component
